requirements.txt
argparse==1.2.1
django-nose==1.4.5
nose==1.2.1
lxml==3.3.5
